requirements.txt
flask>=3.0.0
langchain-openai>=0.1.21
langchain-chroma>=0.1.4
langchain-text-splitters>=0.2.2
langchain-community>=0.3.0
openai>=1.35.0
chromadb>=0.5.5
beautifulsoup4>=4.12.3
requests>=2.32.3
transformers>=4.41.2
torch>=2.3.1

python-dotenv>=1.0.1
gunicorn>=21.2.0
